refactor: use json logging config to allow filter
