+++ jobs/ml_jobs/retrieval_vector/Dockerfile
@@ -16,4 +16,4 @@ EXPOSE 8080
 
 ENV WORKERS=${WORKERS}
 
-CMD hypercorn --bind 0.0.0.0:8080 app.app:app --workers=${WORKERS} --log-config=app/logging.conf
+CMD hypercorn --bind 0.0.0.0:8080 app.app:app --workers=${WORKERS} --log-config="json:app/logging/config.json"
