+++ jobs/ml_jobs/retrieval_vector/app/factory/recommendation.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from app.factory.handler import PredictionHandler
 from app.factory.tops import SearchByTopsHandler
-from app.logger import logger
+from app.logging.logger import logger
 from app.models.prediction_request import PredictionRequest
 from app.models.prediction_result import PredictionResult
 from app.retrieval.reco_client import RecoClient
