--- jobs/ml_jobs/retrieval_vector/app/logging.conf
@@ -1,35 +1,0 @@
-[loggers]
-keys=root,hypercorn.error,hypercorn.access
-
-[handlers]
-keys=console_handler
-
-[formatters]
-keys=json_formatter
-
-[logger_root]
-level=INFO
-handlers=console_handler
-
-[logger_hypercorn.error]
-level=INFO
-handlers=console_handler
-qualname=hypercorn.error
-propagate=0
-
-[logger_hypercorn.access]
-level=INFO
-handlers=console_handler
-qualname=hypercorn.access
-propagate=0
-
-[handler_console_handler]
-class=StreamHandler
-level=INFO
-formatter=json_formatter
-args=(sys.stdout,)
-
-[formatter_json_formatter]
-class=pythonjsonlogger.jsonlogger.JsonFormatter
-format=%(name)s %(asctime)s %(levelname)s %(filename)s %(lineno)s %(process)d %(message)s
-rename_fields={"levelname": "severity", "asctime": "timestamp"}
